Fixed login for users
Notice my comment for checking if the username or password
is empty. I left it in the PHP code for now. For the next
assignment, this will be done on the client side using Javascript

Added PHP prepared statements

A person now can login

If they are a user the program displays Welcome Back ....
else it displays they are not a user.

If the password for the user is incorrect they will get
notified