Fixed the color schemes for the index HTML. Put the menu bar inside the main_div. Removed "margin-top" on the main_div because it was giving extra space and the vertical bar wasn't at the very top.
Removed the body in header.css and put it into the index.css since the index.html has reference to both css files.

Changed the id name "filler-text" to "welcome-text". Changed the color from gray to blue to match the book color on the alien logo. Made the text bold, so it can stand out more.

Made the hover on the menu bar to match the color of the welcome text. The vertical bar menu hover has the same color scheme as the menu bar.

The vertical bar has white text now, and shifted the vertical menu bar over a few pixels.